section_5/requirements.txt
lxml==4.1.1
openpyxl==2.4.9
PyPDF2==1.26.0
python-docx==0.8.6
PyYAML==3.12